update tests to keep in sync
